[tools/depends] Python rework
Currently when crosscompiling python, there are flags that leak into a target build of python.
This implements a clean build of both host and target, and implements the use of crossenv
to insure no buildhost flags are passed through when modules are cross compiled.

This allows us to remove patches for modules that affect distutils/setuptools.
Additional modules can now be quite easily added by building any dependencies, and using the
crossenv to build the modules.

remove copied cross-file.meson when distclean ran