[cmake] macros create core_optional_package_lib function
This function allows to package a TARGET based library that is not
used as part of the build, or linking steps of the core app.

The function sets a variable ``package_libs`` made up of the target names
of libs to be packaged for a platform. The packaging is to be handled by the
platforms packaging system.

Add package_libs dependency to main app target. any potential libs using
this are not linked to the project, and only need to exist before any
packaging steps may occur